requirements.txt
argparse
asyncio
asyncqt
asyncslot
autopep8
cerberus
customtkinter
easydict
empy
flake8-blind-except
flake8-builtins
flake8-class-newline
flake8-comprehensions
flake8-deprecated
flake8-import-order
flake8-quotes
geocoder
gpxpy
grpcio-tools
imutils
jaraco.collections
jsonschema
kconfiglib
kornia
mavsdk
numpydoc
nunavut
opencv-contrib-python
opencv-python
orange3
pip-chill
pkgconfig
protoc-gen-mavsdk
pydub
pyinstaller
pymavlink
pyperclip
pyqt5-tools
pyqt6
pyqtwebengine
pyros-genmsg
pyserial
pyshine
pyside2
pytest-repeat
pytest-rerunfailures
pyulog
s3cmd
serial
shapely
toml
torchvision
tqdm
